backend/requirements-additional.txt
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests

